[step4] refactor Lotto 공통 클래스 만들기
