+++ src/main/java/lotto/AutoLotto.java
@@ -1,31 +1,10 @@
 package lotto;
 
-import java.util.Collections;
-import java.util.List;
+import java.util.Set;
 
-public class AutoLotto {
+public class AutoLotto extends Lotto {
 
-  private final List<LottoNo> lotto;
-
-  public AutoLotto(List<LottoNo> lotto) {
-    this.lotto = lotto;
-    Collections.sort(this.lotto);
-  }
-
-  public List<LottoNo> getLotto() {
-    return lotto;
-  }
-
-  @Override
-  public String toString() {
-    return lotto.toString();
-  }
-
-  public boolean has(BonusBall bonusBall) {
-    return lotto.stream().anyMatch(bonusBall::equals);
-  }
-
-  public int countMatchesNumber(WinningNumbers winningNumbers) {
-    return (int) lotto.stream().filter(winningNumbers::has).count();
+  public AutoLotto(Set<LottoNo> lotto) {
+    super(lotto);
   }
 }
